New panel

+++ src/Components/ProjectOne/ProjectMain/ProjectPanel/ProjectPanel.js
@@ -8,6 +8,10 @@ const ProjectPanel = ({ logged, onLogout }) => {
     const [login, setLogin] = useState('');
     const [password, setPassword] = useState('');
     const [userId, setUserId] = useState('');
+    const [showDelete, setShowDelete] = useState(false);
+    const [showAddUser, setShowAddUser] = useState(false);
+    const [emptyInputError, setEmptyInputError] = useState(false);
+    const [emptyInputErrorDel, setEmptyInputErrorDel] = useState(false);
 
 
     useEffect(() => {
@@ -25,39 +29,55 @@ const ProjectPanel = ({ logged, onLogout }) => {
     }, [userList]);
 
     const handleCheckCredentials = () => {
+        if (!login || !password) {
+            setEmptyInputError(true);
+            return;
+          }
         writeToDatabase(login, password)
         setLogin('');
         setPassword('');
+        setEmptyInputError(false);
 
     };
 
     const handleDeletion = () => {
+        if (!userId) {
+            setEmptyInputErrorDel(true);
+            setUserId('');
+            return;
+        }
+
         for (const user of userList) {
             if (user.login === userId) {
                 deleteUserFromDatabase(user);
                 setUserId('');
+                setEmptyInputErrorDel(false);
                 break;
-
             }
         }
     };
 
+    const toggleDelete = () => {
+        setShowDelete(!showDelete);
+        if (showDelete) {
+            setUserId('');
+            setEmptyInputErrorDel(false);
+        }
+
+    };
+
+    const toggleAddUser = () => {
+        setShowAddUser(!showAddUser);
+       
+        };
+
     return (
         <div className='project-panel-main'>
             <div className='project-panel-opened'>
-                <div className='container-1'>
-                    <div className='user-list'>
-                        <span>Users</span>
-                        <ul>
-                            {userList.map((user, index) => (
-                                user.login !== logged ? (
-                                    <li key={index}>{user.login}</li>
-                                ) : null
-                            ))}
-                        </ul>
-                    </div>
-                </div>
+
                 <div className='container-2'>
+
+
                     <span>Welcome to the main Panel </span>
                     <span>This small project is to show some basic CRUD operations using Firebase as a database. Real-time database is
                         used to achieve that. The authentication method is a custom one and not one of the given options of firebase.
@@ -65,60 +85,97 @@ const ProjectPanel = ({ logged, onLogout }) => {
 
                     <div className='db-write-del'>
 
-                        <div className='write'><span className='del-write'>Add a user in DB</span>
+                        <div>
+                            <div className='write'>
+                                <div className='login-db fade-in' style={{ display: showAddUser ? 'block' : 'none' }}>
+                                    <input
+                                        type='text'
+                                        placeholder='username'
+                                        onChange={(e) => setLogin(e.target.value)}
+                                        value={login}
+                                        style={{ textAlign: 'left' }}
+                                    />
+                                </div>
+                                <div className='password-db fade-in' style={{ display: showAddUser ? 'block' : 'none' }}>
 
-                            <div className='login-db'>
-                                <label className='write-label'>Username</label>
-                                <input
-                                    type='text'
-                                    onChange={(e) => setLogin(e.target.value)}
-                                    value={login}
+                                    <input
+                                        type='password'
+                                        placeholder='password'
+                                        onChange={(e) => setPassword(e.target.value)}
+                                        value={password}
+                                        style={{ textAlign: 'left' }}
+                                    />
+                                </div>
+                                {emptyInputError && <div><span className="write-msg fade-in">Input cannot be empty. Please try again.</span></div>}
+                                <div className='main-button'>
+                                    {showAddUser && (
+                                        <button className='main-btn fade-infade-in' onClick={handleCheckCredentials}>
+                                            Confirm
+                                        </button>
+                                    )}
+                                    <button className='main-btn fade-in' onClick={toggleAddUser}>
+                                        {showAddUser ? 'Cancel' : 'Add'}
+                                    </button>
 
-                                />
+                                </div>
                             </div>
-                            <div className='password-db'>
-                                <label className='write-label'>Password</label>
-                                <input
-                                    type='password'
-                                    onChange={(e) => setPassword(e.target.value)}
-                                    value={password}
-                                />
-                            </div>
-
-                            <div className='main-button'>
-                                <button className='main-btn' onClick={handleCheckCredentials}>
-                                    Add
-                                </button>
-
-                            </div>
-
                         </div>
 
-                        <div className='del'> <span className='del-write'>Delete a user</span>
-                            <div className='remove-db'>
-                                <label className='write-label'>Username</label>
+                        <div className='main-button'>
+                            {showDelete && (<div className='remove-db fade-in'>
+
                                 <input
                                     type='text'
                                     onChange={(e) => setUserId(e.target.value)}
                                     value={userId}
+                                    style={{ textAlign: 'left' }}
                                 />
-                            </div>
-
-                            <div className='main-button'>
-                                <button className='main-btn' onClick={handleDeletion}>
-                                    Delete
+                            </div>)}
+                            {emptyInputErrorDel && <div> <span className='del-msg fade-in '>Wrong Username</span></div>}
+                            {showDelete && (
+                                <button className='main-btn fade-in' onClick={handleDeletion}>
+                                    Confirm
                                 </button>
 
-                            </div>
+                            )}
+                            <button className='main-btn fade-in' onClick={toggleDelete}>
+                                {showDelete ? 'Cancel' : 'Delete'}
+                            </button>
 
 
                         </div>
 
                     </div>
 
+                </div>
 
+                <div className='container-1'>
+                    <div className='user-list'>
+                        <span>User List</span>
+                        <table className='tableData'>
 
+                            <thead>
+                                <tr>
+                                    <th style={{ textAlign: 'start' }}>{userList.map((user, index) => (
+                                        user.login !== logged ? (
+                                            <li className='fade-in' key={index}>{user.login}</li>
+                                        ) : null
+                                    ))}</th>
+                                    <th style={{ textAlign: 'start' }}>{userList.map((user, index) => (
+                                        user.login !== logged ? (
+                                            <li className='fade-in' key={index}>{user.password}</li>
+                                        ) : null
+                                    ))}</th>
+                                    <th style={{ textAlign: 'start' }}>{userList.map((user, index) => (
+                                        user.login !== logged ? (
+                                            <li className='fade-in'> Live!</li>
+                                        ) : null
+                                    ))}</th>
+                                </tr>
+                            </thead>
 
+                        </table>
+                    </div>
                 </div>
                 <div className='container-3'>
                     <div className='logged-in'>Logged: <span>{logged}</span></div>
@@ -128,6 +185,8 @@ const ProjectPanel = ({ logged, onLogout }) => {
                         </button>
                     </div>
                 </div>
+
+
             </div>
         </div>
     );
